requirements.txt
# Core
numpy>=1.24.0
pandas>=2.0.0

# Dataset access
kagglehub>=0.3.0

# Image processing
opencv-python>=4.8.0
Pillow>=10.0.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0

# Jupyter
jupyterlab>=4.0.0
ipywidgets>=8.0.0

# Deep Learning
# NOTE: The default pip install gives you CPU-only PyTorch.
# For NVIDIA GPU (CUDA) support on Windows/Linux, install separately:
#   pip install torch torchvision --index-url https://download.pytorch.org/whl/cu128
# For Mac (Apple Silicon / MPS) or CPU-only, the default works:
#   pip install torch torchvision
torch>=2.0.0
torchvision>=0.15.0

# Object detection (Stage 1)
ultralytics>=8.0.0

# OCR (Stage 2)
pytesseract>=0.3.10
easyocr>=1.7.0

# Knowledge graph (Stage 3)
networkx>=3.1
rdflib>=6.3.0
neo4j>=5.20.0

# Query engine / LLM (Stage 4 — Ollama via HTTP)
requests>=2.28.0

# UI (Stage 5)
streamlit>=1.28.0

# XML/YAML parsing
lxml>=4.9.0
pyyaml>=6.0.0

# Development
pytest>=7.4.0
